Improve structure of project.
The structure is more modular and organized now. Each type of command
is now placed in it's own respective `.cpp` file under the new `commands`
directory. Consequently, the build system has been updated too.
