+++ src/luna.cpp
@@ -1,11 +1,11 @@
-#include "shell.hpp"
-#include "../tests/test.hpp"
+#include "../include/shell.hpp"
+#include "../include/test.hpp"
 
 int main() {
     if (getenv("LUNA_TESTING_ON")) {
         Test test;
         return 0; // no running shell after tests
     }
-    Shell moon;
+    Shell luna;
     return 0;
 }
